23/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
packaging==22.0
Pillow==9.3.0
pytesseract==0.3.10
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13
